requirements.txt
# Python dependencies
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic>=2.3.0
jinja2>=3.1.2
python-dotenv>=1.0.0
python-multipart>=0.0.6
google-genai>=0.3.0

# Testing dependencies
pytest>=7.4.0
httpx>=0.24.1
beautifulsoup4>=4.12.2